move LocalHub to @internal in core-backend
