+++ core/backend/src/test/index.ts
@@ -5,7 +5,6 @@
 export * from "./HubMock";
 export * from "./IModelTestUtils";
 export * from "./KnownTestLocations";
-export * from "./LocalHub";
 export * from "./RevisionUtility";
 export * from "./TestChangeSetUtility";
 export * from "./TestUtils";
